Split event store into modules
